feat: 캐러셀을 공통컴포넌트로 분리
